+++ UVtools.Core/FileFormats/PhotonWorkshopFile.cs
@@ -43,7 +43,7 @@ namespace UVtools.Core.FileFormats
             0
         };
 
-        // CRC-16-ANSI (aka CRC-16-IMB) Polynomial: x^16 + x^15 + x^2 + 1
+        // CRC-16-ANSI (aka CRC-16-IBM) Polynomial: x^16 + x^15 + x^2 + 1
         public static readonly int[] CRC16Table = {
             0x0000, 0xc0c1, 0xc181, 0x0140, 0xc301, 0x03c0, 0x0280, 0xc241,
             0xc601, 0x06c0, 0x0780, 0xc741, 0x0500, 0xc5c1, 0xc481, 0x0440,